Fix equipping item bug
